HTBHF-1845 Renamed all card address files except those without content changes.
