+++ src/test/common/steps/pages.js
@@ -3,7 +3,7 @@ const EnterNino = require('../page/enter-nino')
 const Overview = require('../page/overview')
 const EnterDOB = require('../page/enter-dob')
 const AreYouPregnant = require('../page/are-you-pregnant')
-const Address = require('../page/card-address')
+const Address = require('../page/address')
 const PhoneNumber = require('../page/phone-number')
 const Check = require('../page/check')
 const Confirm = require('../page/confirm')
