+++ src/web/routes/application/card-address/card-address.test.js
@@ -1,6 +1,6 @@
 const test = require('tape')
 const { assocPath } = require('ramda')
-const { contentSummary } = require('./card-address')
+const { contentSummary } = require('./address')
 
 const req = {
   t: string => string,
